refactor: request validation (#29)
* refactor: use service methods in request validation

* refactor: add cancellation token to service methods

* refactor: use cancellation token in validators

* tests: remove redundant test and format code

* refactor: replace ToUpperInvariant with ToUpper

* feat: add table for available currencies in account type

* feat: add method for account type currency support check

* refactor: change names of lambda arguments in validators

+++ BankApp/Entities/Country.cs
@@ -6,7 +6,7 @@ namespace BankApp.Entities;
 [Table("Countries")]
 public class Country
 {
-    [Key] public short Id { get; set; }
+    [Key] public int Id { get; set; }
     public string Code { get; set; } = default!;
     public string Name { get; set; } = default!;
 }
